+++ src/extensions/plugin/hooks/use-view-model.ts
@@ -8,7 +8,7 @@ import { MutableRefObject, useEffect, useRef } from 'react';
 import { DefaultPluginSpec, PluginSpec } from '../../../mol-plugin/spec';
 import { PluginViewModel } from '../view-model';
 
-export function useCreatePluginViewModel(options?: { spec?: PluginSpec | ((defaultSpec?: PluginSpec) => PluginSpec) }): PluginViewModel {
+export function useCreatePluginViewModel(options?: { spec?: PluginSpec | ((defaultSpec: PluginSpec) => PluginSpec) }): PluginViewModel {
     const model = useRef<PluginViewModel>();
     if (!model.current) {
         model.current = new PluginViewModel({